The Quiz App project aims to create an interactive application that allows users to participate in quizzes. Using HTML , CSS & JavaScript
The Quiz App allows users to select a quiz category and answer a set of multiple-choice questions within a specified time limit. After completing , users receive feedback on their answers, including the number of correct and incorrect responses. The app also calculates the overall score and displays it to the user.
The app incorporates features like a countdown timer, progress tracker, and a visually appealing user interface to enhance the user experience. It provides an interactive and engaging platform for users to challenge their knowledge and improve their understanding of different subjects.